+++ packages/jscrambler-cli/src/index.js
@@ -15,7 +15,9 @@ import * as introspection from './introspection';
 import {
   getMatchedFiles,
   validateCustomLabels,
+  validateCodeHardening,
   validateThresholdFn,
+  resolveOutputPath,
 } from './utils';
 
 import getProtectionDefaultFragments, {
@@ -132,6 +134,7 @@ function buildFinalConfig(configPathOrObject) {
 export default {
   Client: JscramblerClient,
   config,
+  resolveOutputPath,
   queries,
   generateSignedParams,
   /**
@@ -320,6 +323,7 @@ export default {
       clientId,
       tolerateMinification,
       codeHardeningThreshold,
+      codeHardening,
       useProfilingData,
       browsers,
       useAppClassification,
@@ -382,6 +386,8 @@ export default {
       throw new Error('Required *filesDest* not provided');
     }
 
+    const normalizedCodeHardening = validateCodeHardening(codeHardening);
+
     let source;
     if (!skipSources) {
 
@@ -426,8 +432,12 @@ export default {
     const updateData = {
       debugMode: !!debugMode,
       tolerateMinification,
-      codeHardeningThreshold
+      codeHardeningThreshold,
     };
+
+    if (typeof normalizedCodeHardening !== 'undefined') {
+      updateData.codeHardening = normalizedCodeHardening;
+    }
 
     if (params && Object.keys(params).length) {
       updateData.parameters = normalizeParameters(params);