+++ packages/jscrambler-metro-plugin/lib/index.js
@@ -398,8 +398,22 @@ module.exports = function (_config = {}, projectRoot = process.cwd()) {
         bundlePaths.bundlePath,
         bundlePaths.bundleSourceMapPath,
       );
-      process.exit(typeof exitCode === 'number' ? exitCode : 0);
-    } catch (err) {
+
+      const isCodeHardeningThresholdSupported =
+        await jscrambler.introspectFieldOnMethod.call(
+          jscrambler,
+          config,
+          'mutation',
+          'createApplicationProtection',
+          'codeHardeningThreshold',
+        );
+
+      // check for incompatible transformations and turn off code hardening
+      handleHermesIncompatibilities(config, isCodeHardeningThresholdSupported);
+
+      // start obfuscation
+      await obfuscateBundle(bundlePath, {fileNames: Array.from(fileNames), entryPointCode}, sourceMapFiles, config, projectRoot);
+    } catch(err) {
       console.error(err);
       process.exit(1);
     }