+++ packages/gulp-jscrambler/CHANGELOG.md
@@ -1,5 +1,20 @@
 # gulp-jscrambler
 
+## 8.5.40
+
+### Patch Changes
+
+- [83c2f6f]: Updated dependencies
+  - jscrambler@8.12.0
+
+## 8.5.39
+
+### Patch Changes
+
+- [25054ca]: Ensure protected files are not written outside of project root folder in gulp
+- [25054ca]: Updated dependencies
+  - jscrambler@8.11.3
+
 ## 8.5.38
 
 ### Patch Changes